repo: normalize line termination across repo
This is a scrub of the entire repo to normalize end-of-line termination to
conform to the recently added .gitattributes file.  A mismatch of line
termination can cause difficult-to-resolve conflicts when rebasing or merging.

In this scrub of the repo, only two files actually required modification.

The process performed is described at:
https://help.github.com/articles/dealing-with-line-endings/#refreshing-a-repository-after-changing-line-endings

The .gitattributes file is required because some files (*.bat and *.sh) are
sensitive to line termination.  The .gitattributes file is stored in the repo,
and overrides settings which could be corrupted by a developer's local
git configuration.  The .gitattributes file was added in f077c81

This commit completes bcc9fec, which fixed a single file.

Change-Id: I5be5c10ae5dd83defe6a05526e4032c5100d9ae9